# api.swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KidsLoop 2.0 REST API",
    "description": "KidsLoop 2.0 backend rest api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:12345/v1",
      "description": "dev"
    },
    {
      "url": "https://cp03fzjc58.execute-api.ap-northeast-2.amazonaws.com/v1",
      "description": "temp test"
    },
    {
      "url": "https://kl2-test.kidsloop.net/v1",
      "description": "test"
    }
  ],
...
  "paths": {
...
    "/schedules/{schedule_id}": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "get schedules by id",
        "description": "get schedules by id",
        "operationId": "GetSchedulesByID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "description": "schedule id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDetailed"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "User not login"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Recorde not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": [
          "schedule"
        ],
        "summary": "update schedule",
        "description": "update schedule",
        "operationId": "updateSchedule",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "description": "schedule id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "allowEmptyValue": false
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "new schedule id",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "string",
                      "description": "new schedule id"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "User not login"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Recorde not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "schedule"
        ],
        "summary": "delete schedule",
        "description": "delete schedule",
        "operationId": "deleteSchedule",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "description": "delete schedule",
            "schema": {
              "type": "string"
            },
            "required": true,
            "allowEmptyValue": false
          },
          {
            "name": "repeat_edit_options",
            "in": "query",
            "description": "repeat edit option",
            "schema": {
              "type": "string",
              "enum": [
                "only_current",
                "with_following"
              ],
              "nullable": true,
              "description": "repeat edit option"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "User not login"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Recorde not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ScheduleUpdate": {
        "type": "object",
        "description": "schedule create info",
        "properties": {
          "title": {
            "type": "string",
            "description": "schedule title"
          },
          "class_id": {
            "type": "string",
            "description": "schedule lessonPlan name"
          },
          "lesson_plan_id": {
            "type": "string",
            "description": "schedule title"
          },
          "teacher_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "schedule lessonPlan name"
          },
          "start_at": {
            "type": "integer",
            "description": "schedule start time, timestamp"
          },
          "end_at": {
            "type": "integer",
            "description": "schedule end time, timestamp"
          },
          "repeat_edit_options": {
            "type": "string",
            "enum": [
              "only_current",
              "with_following"
            ],
            "description": "repeat edit option"
          },
          "repeat": {
            "type": "object",
            "description": "repeat options, only work when mode_type equal Repeat,",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "daily",
                  "weekly",
                  "monthly",
                  "yearly"
                ],
                "description": "repeat type"
              },
              "daily": {
                "type": "object",
                "description": "repeat daily options",
                "properties": {
                  "interval": {
                    "type": "integer",
                    "description": "interval days"
                  },
                  "end": {
                    "$ref": "#/components/schemas/RepeatEnd"
                  }
                }
              },
              "weekly": {
                "type": "object",
                "description": "repeat weekly options",
                "properties": {
                  "interval": {
                    "type": "integer",
                    "description": "interval days"
                  },
                  "on": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "Sunday",
                        "Monday",
                        "Tuesday",
                        "Wednesday",
                        "Thursday",
                        "Friday",
                        "Saturday"
                      ]
                    },
                    "description": "on weekdays"
                  },
                  "end": {
                    "$ref": "#/components/schemas/RepeatEnd"
                  }
                }
              },
              "monthly": {
                "type": "object",
                "description": "repeat monthly options",
                "properties": {
                  "interval": {
                    "type": "integer",
                    "description": "interval months"
                  },
                  "on_type": {
                    "type": "string",
                    "enum": [
                      "date",
                      "week"
                    ],
                    "description": "on type"
                  },
                  "on_date_day": {
                    "type": "integer",
                    "description": "on date day"
                  },
                  "on_week_seq": {
                    "type": "string",
                    "enum": [
                      "first",
                      "second",
                      "third",
                      "fourth",
                      "last"
                    ],
                    "description": "on week seq"
                  },
                  "on_week": {
                    "type": "string",
                    "enum": [
                      "Sunday",
                      "Monday",
                      "Tuesday",
                      "Wednesday",
                      "Thursday",
                      "Friday",
                      "Saturday"
                    ],
                    "description": "on week"
                  },
                  "end": {
                    "$ref": "#/components/schemas/RepeatEnd"
                  }
                }
              },
              "yearly": {
                "type": "object",
                "description": "repeat yearly options",
                "properties": {
                  "interval": {
                    "type": "integer",
                    "description": "interval years (less or equal than 2 years)"
                  },
                  "on_type": {
                    "type": "string",
                    "enum": [
                      "date",
                      "week"
                    ],
                    "description": "on type"
                  },
                  "on_date_month": {
                    "type": "integer",
                    "description": "on date month"
                  },
                  "on_date_day": {
                    "type": "integer",
                    "description": "on date day"
                  },
                  "on_week_month": {
                    "type": "integer",
                    "description": "on week month"
                  },
                  "on_week_seq": {
                    "type": "string",
                    "enum": [
                      "first",
                      "second",
                      "third",
                      "fourth",
                      "last"
                    ],
                    "description": "on week seq"
                  },
                  "on_week": {
                    "type": "string",
                    "enum": [
                      "Sunday",
                      "Monday",
                      "Tuesday",
                      "Wednesday",
                      "Thursday",
                      "Friday",
                      "Saturday"
                    ],
                    "description": "on week"
                  },
                  "end": {
                    "$ref": "#/components/schemas/RepeatEnd"
                  }
                }
              }
            }
          },
          "subject_id": {
            "type": "string",
            "description": "schedule SubjectID"
          },
          "program_id": {
            "type": "string",
            "description": "schedule ProgramID"
          },
          "class_type": {
            "type": "string",
            "enum": [
              "OnlineClass",
              "OfflineClass",
              "Homework",
              "Task"
            ],
            "description": "schedule class_type"
          },
          "due_at": {
            "type": "integer",
            "description": "If I schedule the time as homework or task, then I should be able to set a due date. This is optional for me. The due date cannot be earlier than class end time."
          },
          "description": {
            "type": "string",
            "description": "schedule description"
          },
          "attachment": {
            "$ref": "#/components/schemas/CommonShort"
          },
          "is_force": {
            "type": "boolean",
            "description": "If true, skip the conflict detection"
          },
          "is_repeat": {
            "type": "boolean",
            "description": "If true, schedule will repeat"
          },
          "is_all_day": {
            "type": "boolean",
            "description": "all_day flag"
          }
        }
      },
      "ScheduleDetailed": {
        "type": "object",
        "description": "schedule create info",
        "properties": {
          "id": {
            "type": "string",
            "description": "schedule id"
          },
          "title": {
            "type": "string",
            "description": "schedule title"
          },
          "class": {
            "$ref": "#/components/schemas/CommonShort"
          },
          "lesson_plan": {
            "$ref": "#/components/schemas/CommonShort"
          },
          "teachers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommonShort"
            }
          },
          "start_at": {
            "type": "integer",
            "description": "schedule start time, timestamp"
          },
          "end_at": {
            "type": "integer",
            "description": "schedule end time, timestamp"
          },
          "repeat": {
            "type": "object",
            "description": "repeat options, only work when mode_type equal Repeat,",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "daily",
                  "weekly",
                  "monthly",
                  "yearly"
                ],
                "description": "repeat type"
              },
              "daily": {
                "type": "object",
                "description": "repeat daily options",
                "properties": {
                  "interval": {
                    "type": "integer",
                    "description": "interval days"
                  },
                  "end": {
                    "$ref": "#/components/schemas/RepeatEnd"
                  }
                }
              },
              "weekly": {
                "type": "object",
                "description": "repeat weekly options",
                "properties": {
                  "interval": {
                    "type": "integer",
                    "description": "interval days"
                  },
                  "on": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "Sunday",
                        "Monday",
                        "Tuesday",
                        "Wednesday",
                        "Thursday",
                        "Friday",
                        "Saturday"
                      ]
                    },
                    "description": "on weekdays"
                  },
                  "end": {
                    "$ref": "#/components/schemas/RepeatEnd"
                  }
                }
              },
              "monthly": {
                "type": "object",
                "description": "repeat monthly options",
                "properties": {
                  "interval": {
                    "type": "integer",
                    "description": "interval months"
                  },
                  "on_type": {
                    "type": "string",
                    "enum": [
                      "date",
                      "week"
                    ],
                    "description": "on type"
                  },
                  "on_date_day": {
                    "type": "integer",
                    "description": "on date day"
                  },
                  "on_week_seq": {
                    "type": "string",
                    "enum": [
                      "first",
                      "second",
                      "third",
                      "fourth",
                      "last"
                    ],
                    "description": "on week seq"
                  },
                  "on_week": {
                    "type": "string",
                    "enum": [
                      "Sunday",
                      "Monday",
                      "Tuesday",
                      "Wednesday",
                      "Thursday",
                      "Friday",
                      "Saturday"
                    ],
                    "description": "on week"
                  },
                  "end": {
                    "$ref": "#/components/schemas/RepeatEnd"
                  }
                }
              },
              "yearly": {
                "type": "object",
                "description": "repeat yearly options",
                "properties": {
                  "interval": {
                    "type": "integer",
                    "description": "interval years (less or equal than 2 years)"
                  },
                  "on_type": {
                    "type": "string",
                    "enum": [
                      "date",
                      "week"
                    ],
                    "description": "on type"
                  },
                  "on_date_month": {
                    "type": "integer",
                    "description": "on date month"
                  },
                  "on_date_day": {
                    "type": "integer",
                    "description": "on date day"
                  },
                  "on_week_month": {
                    "type": "integer",
                    "description": "on week month"
                  },
                  "on_week_seq": {
                    "type": "string",
                    "enum": [
                      "first",
                      "second",
                      "third",
                      "fourth",
                      "last"
                    ],
                    "description": "on week seq"
                  },
                  "on_week": {
                    "type": "string",
                    "enum": [
                      "Sunday",
                      "Monday",
                      "Tuesday",
                      "Wednesday",
                      "Thursday",
                      "Friday",
                      "Saturday"
                    ],
                    "description": "on week"
                  },
                  "end": {
                    "$ref": "#/components/schemas/RepeatEnd"
                  }
                }
              }
            }
          },
          "subject": {
            "$ref": "#/components/schemas/CommonShort"
          },
          "program": {
            "$ref": "#/components/schemas/CommonShort"
          },
          "class_type": {
            "type": "string",
            "enum": [
              "OnlineClass",
              "OfflineClass",
              "Homework",
              "Task"
            ],
            "description": "schedule class_type"
          },
          "due_at": {
            "type": "integer",
            "description": "If I schedule the time as homework or task, then I should be able to set a due date. This is optional for me. The due date cannot be earlier than class end time."
          },
          "description": {
            "type": "string",
            "description": "schedule description"
          },
          "attachment": {
            "$ref": "#/components/schemas/CommonShort"
          },
          "is_all_day": {
            "type": "boolean",
            "description": "all_day flag"
          }
        }
      },
      "CommonShort": {
        "type": "object",
        "description": "short info",
        "properties": {
          "id": {
            "type": "string",
            "description": "id"
          },
          "name": {
            "type": "string",
            "description": "name"
          }
        }
      },
      "RepeatEnd": {
        "type": "object",
        "description": "repeat end options",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "never",
              "after_count",
              "after_time"
            ]
          },
          "after_count": {
            "type": "integer",
            "description": "end after N count"
          },
          "after_time": {
            "type": "integer",
            "description": "end after the time(unix timestamp)"
          }
        }
      }
    }
  }
}